+++ reserver/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 """Reserver modules."""
-from .reserver_param import RESERVER_VERSION
-from .reserver_obj import PyPIUploader
-from .reserver_errors import ReserverBaseError
+from .params import RESERVER_VERSION
+from .uploader import PyPIUploader
+from .errors import ReserverBaseError
 
 __version__ = RESERVER_VERSION
